- Introduced a new API endpoint: `/schema/job`.   - Serves the JSON schema for the `ScheduledJob` struct.   - Dynamically generates the schema using `jsonschema.Reflector` to ensure accuracy and consistency. 📋 - Added the `GetJobSchema` handler in the `scheduler` package for serving the schema.
